+++ etils/edc/__init__.py
@@ -15,5 +15,6 @@
 """Dataclasses utils."""
 
 from etils.edc.cast_utils import AutoCast
+from etils.edc.dataclass_utils import __repr__ as repr  # pylint: disable=redefined-builtin
 from etils.edc.dataclass_utils import dataclass
 from etils.edc.field_utils import field