+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog follow https://keepachangelog.com/ format.
         `concurrent` also trigger `concurrent.futures` import).
 *   `eapp` (Added):
     *   Added: `.make_flags_parser` to define CLI flags through dataclasses.
+    *   Added: `.better_logging` to display logs by default, tqdm compatibility,...
 *   `edc`:
     *   Changed: `.replace` only added once (not in subclasses).