+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Changelog follow https://keepachangelog.com/ format.
     *   Added: `ecolab.patch_graphviz` to fix Colab display with `graphviz`.
     *   Added: `ecolab.set_verbose` to activate stderr logging on Colab.
     *   Changed: `ecolab.clear_cached_modules` accept single `str`
+    *   Changed: `ecolab.clear_cached_modules` has a `invalidate=False` to not
+        invalidate previous instances of the modules.
 *   `edc`:
     *   Added: Expose `edc.repr`, for functional use or directly assign class
         members (e.g. `__repr__ = edc.repr`)
@@ -163,7 +165,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 *   Automated github release
 
-[Unreleased]: https://github.com/google/etils/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/google/etils/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/google/etils/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/google/etils/compare/v0.7.0...v0.8.0
 [0.7.1]: https://github.com/google/etils/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/google/etils/compare/v0.6.0...v0.7.0